requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
streamlit>=1.35.0
pyyaml>=6.0
